flatten snapshot service facade into index.ts
